requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
chardet==4.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1 ; python_version >= '3.6'
django-address==0.2.5
django-dotenv==1.4.2
django-phonenumber-field[phonenumberslite]==5.2.0
django-widget-tweaks==1.4.8
django==2.2.5
idna==2.10 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumberslite==8.12.25
pillow==8.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1 ; python_version >= '3.5'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
psycopg2==2.8
